refactor(shielded): address PR #1730 review feedback
- Extract mint/melt verification into a dedicated MintMeltVerifier, wired
  into VertexVerifiers (analogous to NanoHeaderVerifier).
- Fix get_header_sighash_bytes: handle ShieldedOutputsHeader/
  UnshieldBalanceHeader explicitly instead of hitting the unreachable guard.
- hathorlib mint/melt headers: add slots/kw_only, hoist MAX_AMOUNT constant,
  parenthesize range checks, and stub the unused VertexBaseHeader methods
  with NotImplementedError (real path is the framework free functions).
- Drop the redundant except clause in _mint_melt_header deserialization.
- Adapt tests to the free-function (de)serialization path.
